requirements.txt
flask
PyMuPDF
python-docx
sentence-transformers
scikit-learn
pandas
torch
transformers
